+++ src/git_cache/commands/ls_remote.py
@@ -58,8 +58,7 @@ def git_ls_remote(git_options):
     if repository is None or repository == "origin":
         mirror_url = get_mirror_url(git_options)
 
-    supported_prefixes = ["http://", "https://", "ssh://"]
-    if repository and any(repository.startswith(prefix) for prefix in supported_prefixes):
+    elif repository and GitMirror.get_mirror_path(repository):
         if use_mirror_for_remote_url(repository):
             mirror_url = repository